requirements.txt
# Install order matters! 
# Last updated: 2024-01-15 17:00 - Using pre-requirements.txt for any-agent extras

# Step 1: Core dependencies
python-dotenv
pydantic>=2.0.0
typing-extensions
nest-asyncio
aiohttp

# Step 2: Install a2a-sdk (provides 'a2a' module)
# Pin to specific version for consistency
a2a-sdk==0.2.10

# Step 3: any-agent is now installed via pre-requirements.txt
# This ensures it's installed with the [a2a,openai] extras before other dependencies

# Step 4: Additional dependencies (if not pulled in by any-agent)
# These might already be installed by any-agent[openai] but we list them to be sure
openai>=1.0.0
litellm>=1.0.0

# Step 5: Gradio last (it has many dependencies)
gradio==4.36.0

# Compatibility pins to satisfy transitive dependency constraints discovered during HF Spaces build
markupsafe==2.1.5           # gradio constraint (~=2.0)
fsspec>=2023.1.0,<=2025.3.0 # datasets constraint (<=2025.3.0)

# Optional: HTTP dependencies for lightweight A2A
httpx>=0.24.0
fastapi>=0.115.0
uvicorn>=0.22.0
starlette>=0.46.0,<0.47.0
sse-starlette>=2.3.6,<2.4.0
